--- page-src/agricola/central-board/action-board/숲/index.tsx
@@ -1,36 +1,0 @@
-import { Wood } from '@/shared/resource/wood';
-import styled from '@emotion/styled';
-import { ActionContainer } from 'page-src/agricola/central-board/action-board/shared/components/action-container';
-
-export const 숲 = () => {
-  return (
-    <ActionContainer width={140} height={86} top={203} left={167} title="숲">
-      <ContentWrapper>
-        <Wrapper>
-          <Text>3</Text>
-          <Wood width={15} height={17} />
-        </Wrapper>
-        <Wrapper>
-          <div>누적 n개</div>
-        </Wrapper>
-      </ContentWrapper>
-    </ActionContainer>
-  );
-};
-
-const Wrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 5px;
-`;
-
-const Text = styled.div`
-  font-weight: bold;
-`;
-
-const ContentWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-`;
