+++ page-src/agricola/central-board/central-board.sub/action-board/shared/components/footer.tsx
