+++ pages/agricola.tsx
@@ -1,12 +1,16 @@
 import styled from '@emotion/styled';
 import { CentralBoard } from 'page-src/agricola/central-board';
 import { Header } from 'page-src/agricola/header';
+import { UserSection } from 'page-src/agricola/user-section';
 
 const AgricolaPage = () => {
   return (
     <StyledBackground>
       <Header />
-      <CentralBoard />
+      <BoardWrapper>
+        <CentralBoard />
+        <UserSection />
+      </BoardWrapper>
     </StyledBackground>
   );
 };
@@ -15,6 +19,13 @@ const StyledBackground = styled.div`
   background-image: url('https://x.boardgamearena.net/data/themereleases/current/games/agricola/220107-0030/img/background.jpg');
   height: 100%;
   width: 100%;
+  padding-top: 10px;
+`;
+
+const BoardWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  padding: 0 30px;
   padding-top: 10px;
 `;
